requirements.txt
flask
transformers
torch
PyPDF2
python-docx
sentencepiece
sacremoses
PyMuPDF
